feat: Remove Obsession/MethodOrder cop
Ideally we would all take the time to split refactors out in separate
PRs. In reality, these refactors are always shoved in with the rest of
the changes. This makes it very hard to review even small changes when
methods are moved up and down files -- especially when the body of the
moved methods is changed too.